+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   - unicode plotting extension marked as deprecated (will be removed in next major version), they can be realized now with GridVisualize 1.9 and backend Plotter = UnicdePlots
   - added scalarplot, vectorplot and broken_scalarplot defined for FEVectorBlocks (moved from ExtendableFEM)
   - updated plotting code in all examples
+  - added Curl2D and Curl3D evaluation for Hdiv
+  
 
 ## v1.5.0 November 21, 2025
   - added type parameter to `PointEvaluator` so `lazy_interpolate!` is now accessible for automatic differentiation